[docs] Make homepage docs button point to latest
Previously it pointed to the version list, but this makes it harder
to jump into the docs from the homepage. A user can still click on
Documentation from the nav menu to find the page with all versions or
just remove the version from the URL.

Update intro wording to be more concise.